+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 - added new module for optimization functionalities
 - changed the structure of the `Network.train` method. It is now a high-level method that is used to choose a particular training method
 - recursive least-squares was added as a new training method next to gradient descent
+- added new utility functions for connectivity matrix generation
 
 0.9
 ---